// packages/walletkit/src/clients/tonapi/types/traces.ts
/**
 * Copyright (c) TonTech.
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 *
 */

import type { TonApiTransaction } from './transactions';

export interface TonApiTrace {
    transaction: TonApiTransaction;
    children?: TonApiTrace[];
}